+++ packages/gateway/src/metric/emitter.ts
@@ -9,10 +9,14 @@ import { MetricsChannels, PubSubEvents } from '@graphql-portal/types';
 const logger = prefixLogger('analytics:metric-emitter');
 export const metricEmitter = new EventEmitter();
 
-const lpush = (key: string, ...args: any[]): Promise<number | void> =>
+const lpush = async (key: string, ...args: any[]): Promise<void> => {
+  if (!key) return;
+  if (Object.values(MetricsChannels).includes(key as MetricsChannels) && !args[0]) return;
+  logger.debug(`lpush: ${key}`);
   redis.lpush(key, ...args).catch((error: Error) => {
     logger.error(`Failed to write request data. \n Error: ${error} \n Data: ${args}`);
   });
+};
 
 const pubSubListenerWrapper = (emit: (...args: any[]) => any) => {
   return (data: { resolverData: ResolverData }): void => {
@@ -24,7 +28,7 @@ const pubSubListenerWrapper = (emit: (...args: any[]) => any) => {
 const subscribe = async (pubsub: MeshPubSub): Promise<void> => {
   if (!metricEmitter.listenerCount(MetricsChannels.GOT_REQUEST)) {
     metricEmitter.on(MetricsChannels.GOT_REQUEST, async (id: string, { query, userAgent, ip, request, date }) => {
-      await lpush(MetricsChannels.REQUEST_IDS, id);
+      logger.debug(`MetricsChannels.GOT_REQUEST ${id}: ${JSON.stringify(query)}`);
       await lpush(
         id,
         serializer(
@@ -42,7 +46,8 @@ const subscribe = async (pubsub: MeshPubSub): Promise<void> => {
       );
     });
 
-    metricEmitter.on(MetricsChannels.RESOLVER_CALLED, (resolverData: ResolverData) =>
+    metricEmitter.on(MetricsChannels.RESOLVER_CALLED, (resolverData: ResolverData) => {
+      logger.debug(`MetricsChannels.RESOLVER_CALLED: ${JSON.stringify(resolverData)}`);
       lpush(
         resolverData.context.requestId,
         serializer({
@@ -50,33 +55,36 @@ const subscribe = async (pubsub: MeshPubSub): Promise<void> => {
           info: resolverData.info,
           args: resolverData.args,
         })
-      )
-    );
+      );
+    });
 
-    metricEmitter.on(MetricsChannels.RESOLVER_DONE, (resolverData: ResolverData, result: any) =>
+    metricEmitter.on(MetricsChannels.RESOLVER_DONE, (resolverData: ResolverData, result: any) => {
+      logger.debug(`MetricsChannels.RESOLVER_DONE: ${JSON.stringify(resolverData)}, ${JSON.stringify(result)}`);
       lpush(
         resolverData.context.requestId,
         serializer({
           ...transformResolverData(MetricsChannels.RESOLVER_DONE, resolverData),
           result,
         })
-      )
-    );
+      );
+    });
 
-    metricEmitter.on(MetricsChannels.RESOLVER_ERROR, (resolverData: ResolverData, error: Error) =>
+    metricEmitter.on(MetricsChannels.RESOLVER_ERROR, (resolverData: ResolverData, error: Error) => {
+      logger.debug(`MetricsChannels.RESOLVER_ERROR: ${JSON.stringify(resolverData)}, ${error}`);
       lpush(
         resolverData.context.requestId,
         serializer({
           ...transformResolverData(MetricsChannels.RESOLVER_ERROR, resolverData),
           error,
         })
-      )
-    );
+      );
+    });
 
     metricEmitter.on(
       MetricsChannels.SENT_RESPONSE,
-      (id: string, rawResponseBody: string, contentLength: number, date: number) =>
-        lpush(
+      async (id: string, rawResponseBody: string, contentLength: number, date: number) => {
+        logger.debug(`MetricsChannels.SENT_RESPONSE: ${id}, ${contentLength}`);
+        await lpush(
           id,
           serializer({
             event: MetricsChannels.SENT_RESPONSE,
@@ -84,30 +92,36 @@ const subscribe = async (pubsub: MeshPubSub): Promise<void> => {
             contentLength,
             date,
           })
-        )
+        );
+        await lpush(MetricsChannels.REQUEST_IDS, id);
+      }
     );
 
-    metricEmitter.on(MetricsChannels.GOT_ERROR, (id: string, error: Error, date: number) =>
-      lpush(
+    metricEmitter.on(MetricsChannels.GOT_ERROR, async (id: string, error: Error, date: number) => {
+      logger.debug(`MetricsChannels.GOT_ERROR: ${id}, ${error}`);
+      await lpush(
         id,
         serializer({
           event: MetricsChannels.GOT_ERROR,
           error,
           date,
         })
-      )
-    );
+      );
+      await lpush(MetricsChannels.REQUEST_IDS, id);
+    });
   }
 
   await pubsub.subscribe(
     PubSubEvents.RESOLVER_CALLED,
     pubSubListenerWrapper(({ resolverData }) => {
+      logger.debug(`PubSubEvents.RESOLVER_CALLED ${resolverData.context?.id || ''}`);
       metricEmitter.emit(MetricsChannels.RESOLVER_CALLED, resolverData);
     })
   );
   await pubsub.subscribe(
     PubSubEvents.RESOLVER_DONE,
     pubSubListenerWrapper(({ resolverData, result }) => {
+      logger.debug(`PubSubEvents.RESOLVER_DONE ${resolverData.context?.id || ''}`);
       if (/error/i.test(result?.constructor?.name)) {
         metricEmitter.emit(MetricsChannels.RESOLVER_ERROR, resolverData, result);
       } else {
@@ -118,6 +132,7 @@ const subscribe = async (pubsub: MeshPubSub): Promise<void> => {
   await pubsub.subscribe(
     PubSubEvents.RESOLVER_ERROR,
     pubSubListenerWrapper(({ resolverData, error }) => {
+      logger.debug(`PubSubEvents.RESOLVER_ERROR ${resolverData.context?.id || ''}: ${error.message}`);
       metricEmitter.emit(MetricsChannels.RESOLVER_ERROR, resolverData, error);
     })
   );