+++ packages/gateway/src/test/middleware/assign-request-id.spec.ts
@@ -1,10 +1,15 @@
 import { MetricsChannels } from '@graphql-portal/types';
 import { NextFunction, Request, Response } from 'express';
 import assignRequestId from '../../middleware/assign-request-id';
+import { isIntrospectionRequest } from '../../middleware/utils';
 import { metricEmitter } from '../../metric';
 
 jest.mock('uuid', () => ({
   v4: () => 'requestId',
+}));
+
+jest.mock('../../middleware/utils', () => ({
+  isIntrospectionRequest: jest.fn().mockReturnValue(false),
 }));
 
 jest.mock('../../metric/emitter', () => ({
@@ -37,6 +42,7 @@ describe('Assign reequest id MW', () => {
     expect(expressMw).toBeInstanceOf(Function);
 
     expressMw(mockRequest as Request, mockResponse as Response, nextFunction);
+    expect(isIntrospectionRequest).toBeCalled();
     expect(nextFunction).toBeCalled();
     expect(mockRequest.id).toBe('requestId');
     expect(metricEmitter.emit).toBeCalledTimes(1);