+++ packages/datasources/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphql-portal/datasources
 
+## 0.1.0
+
+### Minor Changes
+
+- 6d198a7: fixed custom handlers validation, fixed patches
+
 ## 0.0.3
 
 ### Patch Changes